Running the code uploaded to the UART
Added the code to run the Javascript code.
When the program starts it writes [Waiting for data]

You can start streaming data to it in Intel Hex.
This will decode the data in realtime

When the end of file is being detected
:00000001FF

It will run the javascript parser and run the code
if the parser is OK with it.

The UART output will be redirected so the JerryEditor is able
to output in its console.

Signed-off-by: Sergio Martinez <sergio.martinez.rodriguez@intel.com>